// src/lib/enums/public.ts
export enum CookieEnum { // Cookie
  // Token = 'GFToken', // 会话
  // UserID = 'GFUserID', // 用户ID
  // UserName = 'GFUserName', // 用户名
  // Source = 'GFSource' // 源地址
  Token = 'GFPToken', // 会话
  UserID = 'GFPUserID', // 用户ID
  UserName = 'GFPUserName', // 用户名
  Source = 'GFPSource', // 源地址
  Team = 'GFPTeam' // 团队
}

export enum ParamEnum { // 参数
  // Token = 'gf_token', // 会话
  // UserID = 'gf_userid', // 用户ID
  // UserName = 'gf_username', // 用户名
  // Source = 'gf_source', // 源地址
  // Operate = 'gf_operate' // 操作
  Token = 'gfp_token', // 会话
  UserID = 'gfp_userid', // 用户ID
  UserName = 'gfp_username', // 用户名
  Source = 'gfp_source', // 源地址
  Operate = 'gfp_operate' // 操作
}